+++ lib/plugins/index.js
@@ -49,7 +49,7 @@ module.exports = [
   require('./aws/package/compile/events/cloudWatchLog.js'),
   require('./aws/package/compile/events/cognitoUserPool.js'),
   require('./aws/package/compile/events/eventBridge/index.js'),
-  require('./aws/package/compile/events/sqs/index.js'),
+  require('./aws/package/compile/events/sqs.js'),
   require('./aws/package/compile/events/cloudFront.js'),
   require('./aws/package/compile/events/httpApi'),
   require('./aws/deployFunction.js'),
